+++ frontend/catalyst/jax_primitives.py
@@ -32,6 +32,7 @@ from mlir_quantum.dialects.quantum import (
     CountsOp,
     CustomOp,
     DeallocOp,
+    DeviceOp,
     ExpvalOp,
     ExtractOp,
     HamiltonianOp,
@@ -70,12 +71,10 @@ class Qbit:
         self.aval = AbstractQbit()
 
 
-# pylint: disable=abstract-method
 class AbstractQbit(jax.core.AbstractValue):
     """Abstract Qbit"""
 
 
-# pylint: disable=abstract-method
 class ConcreteQbit(AbstractQbit):
     """Concrete Qbit."""
 
@@ -95,12 +94,10 @@ class Qreg:
         self.aval = AbstractQreg()
 
 
-# pylint: disable=abstract-method
 class AbstractQreg(jax.core.AbstractValue):
     """Abstract quantum register."""
 
 
-# pylint: disable=abstract-method
 class ConcreteQreg(AbstractQreg):
     """Concrete quantum register."""
 
@@ -120,7 +117,6 @@ class Obs:
         self.aval = AbstractObs(num_qubits, primitive)
 
 
-# pylint: disable=abstract-method
 class AbstractObs(jax.core.AbstractValue):
     """Abstract observable."""
 
@@ -129,7 +125,6 @@ class AbstractObs(jax.core.AbstractValue):
         self.primitive = primitive
 
 
-# pylint: disable=abstract-method
 class ConcreteObs(AbstractObs):
     """Concrete observable."""
 
@@ -159,6 +154,8 @@ mlir.ir_type_handlers[AbstractObs] = _obs_lowering
 # Primitives #
 ##############
 
+qdevice_p = jax.core.Primitive("qdevice")
+qdevice_p.multiple_results = True
 qalloc_p = jax.core.Primitive("qalloc")
 qdealloc_p = jax.core.Primitive("qdealloc")
 qdealloc_p.multiple_results = True
@@ -256,8 +253,6 @@ def _func_lowering(ctx, *args, call_jaxpr, fn, call=True):
 #
 # grad
 #
-
-
 @grad_p.def_impl
 def _grad_def_impl(ctx, *args, jaxpr, fn, method, h, argnum):  # pragma: no cover
     raise NotImplementedError()
@@ -301,6 +296,7 @@ def _grad_lowering(ctx, *args, jaxpr, fn, method, h, argnum):
     flat_output_types = util.flatten(output_types)
     constants = [ConstantOp(ir.DenseElementsAttr.get(const)).results for const in jaxpr.consts]
     args_and_consts = constants + list(args)
+
     return GradOp(
         flat_output_types,
         ir.StringAttr.get(method),
@@ -309,6 +305,37 @@ def _grad_lowering(ctx, *args, jaxpr, fn, method, h, argnum):
         diffArgIndices=diffArgIndices,
         finiteDiffParam=finiteDiffParam,
     ).results
+
+
+#
+# qdevice
+#
+def qdevice(spec, val):
+    """Bind operands to operation."""
+    return qdevice_p.bind(spec=spec, val=val)
+
+
+@qdevice_p.def_impl
+def _qdevice_def_impl(ctx, spec, val):  # pragma: no cover
+    raise NotImplementedError()
+
+
+@qdevice_p.def_abstract_eval
+def _qdevice_abstract_eval(spec, val):
+    return ()
+
+
+def _qdevice_lowering(jax_ctx: mlir.LoweringRuleContext, spec, val):
+    ctx = jax_ctx.module_context.context
+    ctx.allow_unregistered_dialects = True
+
+    backend_attr = ir.StringAttr.get(spec)
+    backend_val = "default" if val == "qjit.device" else val
+    val_attr = ir.StringAttr.get(backend_val)
+
+    DeviceOp(specs=ir.ArrayAttr.get([backend_attr, val_attr]))
+
+    return ()
 
 
 #
@@ -374,8 +401,6 @@ def _qdealloc_lowering(jax_ctx: mlir.LoweringRuleContext, qreg):
 #
 # qextract
 #
-
-
 @qextract_p.def_impl
 def _qextract_def_impl(ctx, qreg, qubit_idx):  # pragma: no cover
     raise NotImplementedError()
@@ -1272,8 +1297,9 @@ def _qfor_lowering(
 #
 # registration
 #
-mlir.register_lowering(qdealloc_p, _qdealloc_lowering)
+mlir.register_lowering(qdevice_p, _qdevice_lowering)
 mlir.register_lowering(qalloc_p, _qalloc_lowering)
+mlir.register_lowering(qdealloc_p, _qdealloc_lowering)
 mlir.register_lowering(qextract_p, _qextract_lowering)
 mlir.register_lowering(qinsert_p, _qinsert_lowering)
 mlir.register_lowering(qinst_p, _qinst_lowering)